requirements.txt
flask
flask-mongoengine
nose
Flask-Bootstrap
Flask-Security
Flask-Login==0.2.11
Flask-Admin
flask-debugtoolbar
flask-cache
pymongo==3.5.1
flask_simple_sitemap
awesome-slugify
